BioSemi BDF

EEGIO.BDF — Method

BDF(data::Array, sRate::Integer; path="", file="", kwargs...)

Covenience constructor for BDF struct. To get a valid BDF struct, you need to provide the data and its sampling rate. All other keyword arguments are optional and will be passed to the BDFHeader constructor. Although optional, the more information you provide, the less likely you will face problems when importing the data into other software.

Arguments

data::Array: 2D matrix of data. Each row represents a sample and each column a channel.

sRate::Integer: Sampling rate of the data in Hz.

Keyword arguments

path::String: Path to the folder where the file will be saved. If not provided, you will

have to provide the full path when calling the write_bdf function.

file::String: Name of the file. If not provided, you will have to provide in the path.

subID::String: Subject ID.

recID::String: Recording ID.

startDate::String: Date of the recording.

startTime::String: Time of the recording.

recordDuration::Integer: Duration of each data record in seconds.

chanLabels::Vector{String}: Channel labels.

transducer::Vector{String}: Transducer type, eg. 'active electrode'.

physDim::Vector{String}: Physical dimension of the data, eg. 'uV'.

physMin::Vector{Int}: Minimum physical value of the data.

physMax::Vector{Int}: Maximum physical value of the data.

digMin::Vector{Int}: Minimum digital value of the data.

digMax::Vector{Int}: Maximum digital value of the data.

prefilt::Vector{String}: Prefiltering of the data, eg. 'HP:0.1Hz LP:75Hz'.

reserved::Vector{String}: Reserved fields (usually empty).

Returns

BDF: BDF struct.

Additionally, this constructor provides validation checks that might help you avoid mistakes or getting the data distorted during conversion to Int24 while writing to disk.

Examples

# Create a BDF struct with 10s of random data and 20 channels sampled at 256Hz.
BDF(rand(2560, 20), 256)

If you already have a BDFHeader, you can use the default constructor:

BDF(header, data, path="path/to/file/", file="file.bdf")

EEGIO.BDFHeader — Method

BDFHeader(samples::Integer, channels::Integer, sRate::Integer; kwargs...)

Convenience constructor for the BDFHeader struct. The only three necessary arguments is the number of samples in the data, the number of channels, and the sampling rate of signals. Other fields are optional, but providing them might improve compatibility with other software.

Arguments

samples::Integer: Number of samples in the data.

channels::Integer: Number of channels in the data.

sRate::Integer: Sampling rate of the data.

Keyword arguments

subID::String: Subject ID.

recID::String: Recording ID.

startDate::String: Date of the recording.

startTime::String: Time of the recording.

recordDuration::Integer: Duration of each data record in seconds.

chanLabels::Vector{String}: Channel labels.

transducer::Vector{String}: Transducer type, eg. 'active electrode'.

physDim::Vector{String}: Physical dimension of the data, eg. 'uV'.

physMin::Vector{Int}: Minimum physical value of the data.

physMax::Vector{Int}: Maximum physical value of the data.

digMin::Vector{Int}: Minimum digital value of the data.

digMax::Vector{Int}: Maximum digital value of the data.

prefilt::Vector{String}: Prefiltering of the data, eg. 'HP:0.1Hz LP:75Hz'.

reserved::Vector{String}: Reserved fields (usually empty).

Returns

BDFHeader: A BDFHeader struct with the provided values.

You might notice that couple of fields mentioned in the BDF specification are not included in the constructor. These either have a fixed value or are calculated from provided values to reduce the risk of mistakes. If you want to change these values, you can overwrite the fields of the returned struct directly or use the default constructor and fill all the fields manually. However, this constructor provides additional validation checks that might help you avoid entering wrong values or getting the data distorted during conversion to Int24 while writing to disk.

Examples

header = BDFHeader(1024, 17, 256)

header = BDFHeader(1024, 17, 256, subID="Subject 1", recID="Recording 1", startDate="01.01.2019", startTime="00:00:00")

EEGIO.read_bdf — Method

read_bdf(f::String; kwargs...)

Read data from a BDF file.

Providing a string containing valid path to a .bdf file will result in reading the data as a Float64 matrix with an added offset. Behavior of the function can be altered through additional keyword arguments listed below. Please consult the online documentation for a more thorough explanation of different options.

Arguments:

f::String

Path to the BDF file to be read.

Keywords:

onlyHeader::Bool=false

Indicates whether to read the header information with data points or just the header.

addOffset::Bool=true

Whether to add constant offset to data points to compensate for asymmetric digitalization done by the amplifier. This setting differs between software. Defaults to true.

numPrecision::Type=Float64

Specifies the numerical type of the data. Data points in BDF files are stored as signed 24-bit integers, therefore can be read as types with higher bit count per number. Possible tested options: Float32, Float64, Int32, Int64. Since there is no clear benefit (except amount of memory used) to using smaller number, the default is set to Float64.

chanSelect::Union{Int, Range, Vector{Int}, String, Vector{String}, Regex, Symbol}=:All

Specifies the subset of channels to read from the data. Depending on the provided value you can select different subsets of channels. Using integers (either single number, range, or a vector) will select channels by their position in the data (e.g. chanSelect=[1,4,8], will pick the first, fourth, and eighth). Using strings or regex expression will pick all the channels with matching names stored in the header. Finally, you can provide symbols :None or :All to read neither or every channel available. Default is set to :All.

chanIgnore::Union{Int, UnitRange, Vector{Int}, String, Vector{String}, Regex, Symbol}=:None

Specifies the subset of channel to omit while reading the data. Uses the same selectors as chanSelect and picked values are subtracted from the set of electrodes chosen by chanSelect. Default is set to :None.

timeSelect::Union{Int, UnitRange, Tuple{AbstractFloat}, Symbol}=:All

Specifies the part of the time course of the data to be read. BDF files are read one record at a time (just as they are written), so the user can indicate which records to read. Using integers will select records with those indexes in the data. Using a tuple of floats will be interpreted as start and stop values in seconds and all records that fit this time span will be read. Using Symbol :All will read every record in the file. This is also the default.

readStatus::Bool=true

Specifies if the Status channel should be read. As this is a special channel, always placed at the end of the file and carrying trigger information along with some technical data about the hardware setup, this setting provides an option to ignore it. Please be aware, that its absence might generate errors in code that expects it (even if it is empty). Default is set to true.

Current implementation follows closely the specification formulated by BioSemi that extends EDF format from 16 to 24-bit. Therefore it will read only data produced by BioSemi equipment and software compliant with the specification. BDF+ extension is not yet implemented.

EEGIO.read_bdf_header — Method

read_bdf_header(::IO)

Read the header of a BDF file.

BrainVision EEG

EEGIO.read_eeg — Method

read_eeg(f::String; kwargs...)

Read data from an EEG file.

Providing a string containing valid path to a .eeg, .vhdr, or .vmrk file will result in reading the data as a Float64 matrix (provided all files share their name). Behavior of the function can be altered through additional keyword arguments listed below. Please consult the online documentation for a more thorough explanation of different options.

Arguments:

f::String

Path to the EEG file (either .eeg, .vhdr, or .vmrk) to be read.

Keywords:

onlyHeader::Bool=false

Indicates whether to read the header information with data points or just the header.

onlyMarkers::Bool=false

Indicates whether to read the data points or just the header and markers.

numPrecision::Type=Float64

Specifies the numerical type of the data. Data points in EEG files are stored as 16-bit integers or 32-bit floats, therefore can be read as types with higher bit count per number. Possible tested options: Float32, Float64, Int32, Int64. Since there is no clear benefit (except amount of memory used) to using smaller number, the default is set to Float64.

chanSelect::Union{Int, Range, Vector{Int}, String, Vector{String}, Regex, Symbol}=:All

Specifies the subset of channels to read from the data. Depending on the provided value you can select different subsets of channels. Using integers (either single number, range, or a vector) will select channels by their position in the data (e.g. chanSelect=[1,4,8], will pick the first, fourth, and eighth). Using strings or regex expression will pick all the channels with matching names stored in the header. Finally, you can provide symbols :None or :All to read neither or every channel available. Default is set to :All.

chanIgnore::Union{Int, UnitRange, Vector{Int}, String, Vector{String}, Regex, Symbol}=:None

Specifies the subset of channel to omit while reading the data. Uses the same selectors as chanSelect and picked values are subtracted from the set of electrodes chosen by chanSelect. Default is set to :None.

timeSelect::Union{Int, UnitRange, Tuple{AbstractFloat}, Symbol}=:All

Specifies the part of the time course of the data to be read. Using integers will select samples with those indexes in the data. Using a tuple of floats will be interpreted as start and stop values in seconds and all samples that fit this time span will be read. Using Symbol :All will read every sample in the file. This is also the default.

Current implementation follows closely the specification formulated by BrainVision. However the contents of the header and marker files can differ substantially, even if recorded with BrainVision hardware. If your files are not read properly, open an issue on github and try to provide a sample dataset to reproduce the problem.

EEGIO.read_eeg_header — Method

read_eeg_header(::String)

Reader the EEG header based on the name of file.

Utilities

EEGIO.pick_channels — Method

EEGIO.pick_channels(channels, nChannels::Integer, chanLabels::Vector{String})

Internal function used to properly select requested subset of channels from the data. It allows users to narrow down the number of read channels through numerical indexes, or name labels picked directly with a list or matched with regex expression.

Always returns a vector of indices corresponding to the picked channels.

Examples

# Assume we have an object `raw` of type BDF with 10 channels.

# Picking the first channel of the data.
pick_channels(1, raw.header.nChannels, raw.header.chanLabels)

# Picking the first 5 channels of the data.
pick_channels(1:5, raw.header.nChannels, raw.header.chanLabels)

# Picking channels Fp1, Fp2, F7, F3, and Fz.
pick_channels(["Fp1", "Fp2", "F7", "F3", "Fz"], raw.header.nChannels, raw.header.chanLabels)

# Picking all channels that contain the letter "F".
pick_channels(r"F", raw.header.nChannels, raw.header.chanLabels)

EEGIO.pick_samples — Method

EEGIO.pick_samples(records, header::BDFHeader)
EEGIO.pick_samples(samples, nDataSamples::Ineger, header::EEGHeader)

Internal function used to properly pick the range of samples that should be read from every channel. Since data is stored differently in each format, specialized variants were written to read as much data as needed to satify the query without compromising efficiency. Please refer to the description of timeSelect keyword argument of particular read function for more detailed account of the picking behaviour.

Always returns a UnitRange.

Examples

# Picking the first 10 records of the data (assuming `file` is an object of type BDF).
pick_samples(1:10, file.header)

# Picking the first 10 seconds of the data (assuming `file` is an object of type EEG).
pick_samples((1.,10.), size(file.data,1), file.header)

Base.split — Method

split(file::EEGData, timepoint)

Allows for splitting the data into two objects at a specified timepoint. timepoint can be either a Float or an Integer. Depending on the format, the timepoint will be measured in different units. To preserve the structure of the underlying format, e.g. BDF files will be split based on the record duration. Returns two new EEGData objects with the split data.

Examples

# Splitting the data at the 10th second of the file.
data1, data2 = split(file, 10.)

# Splitting the data at the 100th sample/record of the file.
data1, data2 = split(file, 100)

EEGIO.crop! — Method

crop!(file::EEGData, timeSelect)

Performs cropping in place. See crop for more details.

EEGIO.crop — Method

crop(file::EEGData, timeSelect)

Allows for cropping the data to a specified time window. timeSelect can be either an Integer, a UnitRange, or a tuple of Floats. Depending on the format, the time window will be measured in different units. Since this keyword argument works exactly the same as in read functions, please refer to their documentation for more details on specific use cases. Returns a new EEGData object with the cropped data.

Examples

# Cropping the data to the first 10 seconds of the file.
data = crop(file, (1.,10.))

# Cropping the data to the first 10 samples/records of the file.
data = crop(file, 1:10)

EEGIO.select! — Method

select!(file::EEGData, chanSelect)

Performs channel selection in place. See select for more details.

EEGIO.select — Method

select(file::EEGData, chanSelect)

Allows for selecting a subset of channels from the original data. chanSelect can be either an Integer, a UnitRange, a Vector of Integers/Strings, or a regular expression. Behaves exactly the same as in read functions. Returns a new EEGData object with only the selected channels.

Examples

# Selecting the first 10 channels from the file.
data = select(file, 1:10)

# Selecting channels 1, 3, 5, and 7 from the file.
data = select(file, [1,3,5,7])

# Selecting only frontal channels (labels containing the letter "F") from the file.
data = select(file, r"F")